+++ packages/tsurlfilter/src/rules/declarative-converter/grouped-rules-converters/remove-header-converter.ts
@@ -2,7 +2,7 @@ import type { IndexedNetworkRuleWithHash } from '../network-indexed-rule-with-ha
 import { type DeclarativeRule } from '../declarative-rule';
 import { type ConvertedRules } from '../converted-result';
 
-import { DeclarativeRuleConverter } from './abstract-rule-converter';
+import { AbstractRuleConverter } from './abstract-rule-converter';
 
 /**
  * Describes how to convert $removeheader rules.
@@ -12,7 +12,7 @@ import { DeclarativeRuleConverter } from './abstract-rule-converter';
  * '$match-case', '$script', '$stylesheet, etc.
  *
  */
-export class RemoveHeaderRulesConverter extends DeclarativeRuleConverter {
+export class RemoveHeaderRulesConverter extends AbstractRuleConverter {
     /**
      * Converts indexed rules grouped by $removeheader into declarative rules:
      * for each rule looks for similar rules and groups them into a new rule.